+++ tests/test_dbentry/test_admin.py
@@ -8,7 +8,6 @@ from django.contrib.admin.views.main import ALL_VAR, ORDER_VAR
 from django.contrib.auth import get_permission_codename
 from django.contrib.auth.models import Permission, User
 from django.contrib.contenttypes.models import ContentType
-from django.contrib.postgres.aggregates import ArrayAgg
 from django.core import exceptions
 from django.db import connections, transaction
 from django.db.models import Count, Exists, Func, Min, Subquery
@@ -42,7 +41,7 @@ class AdminTestMethodsMixin(object):
         return (
             self.queryset
             .filter(pk=obj.pk)
-            .annotate(**self.model_admin.get_changelist_annotations())
+            .annotate(**self.model.get_overview_annotations())
             .get()
         )
 
@@ -237,9 +236,9 @@ class TestAusgabenAdmin(AdminTestMethodsMixin, AdminTestCase):
         """Assert that AusgabenAdmin uses the AusgabeChangeList changelist class."""
         self.assertEqual(self.model_admin.get_changelist(self.get_request()), AusgabeChangeList)
 
-    def test_get_changelist_annotations(self):
-        """Check the annotations for this changelist."""
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('jahr_string', annotations)
         self.assertIsInstance(annotations['jahr_string'], Func)
         self.assertIn('num_string', annotations)
@@ -491,10 +490,11 @@ class TestAutorAdmin(AdminTestMethodsMixin, AdminTestCase):
         cls.obj1 = make(cls.model, magazin__magazin_name=['Testmagazin1', 'Testmagazin2'])
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('magazin_list', annotations)
-        self.assertIsInstance(annotations['magazin_list'], ArrayAgg)
+        self.assertIsInstance(annotations['magazin_list'], Func)
 
     def test_autor_name(self):
         self.assertEqual(self.model_admin.autor_name(self.obj1), self.obj1._name)
@@ -519,16 +519,17 @@ class TestBandAdmin(AdminTestMethodsMixin, AdminTestCase):
         )
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('genre_list', annotations)
-        self.assertIsInstance(annotations['genre_list'], ArrayAgg)
+        self.assertIsInstance(annotations['genre_list'], Func)
         self.assertIn('musiker_list', annotations)
-        self.assertIsInstance(annotations['musiker_list'], ArrayAgg)
+        self.assertIsInstance(annotations['musiker_list'], Func)
         self.assertIn('alias_list', annotations)
-        self.assertIsInstance(annotations['alias_list'], ArrayAgg)
+        self.assertIsInstance(annotations['alias_list'], Func)
         self.assertIn('orte_list', annotations)
-        self.assertIsInstance(annotations['orte_list'], ArrayAgg)
+        self.assertIsInstance(annotations['orte_list'], Func)
 
     def test_genre_string(self):
         obj = self.get_annotated_model_obj(self.obj1)
@@ -685,8 +686,9 @@ class TestBaseBrochureAdmin(AdminTestCase):
         self.assertIn('jahr_min', annotations)
         self.assertEqual(annotations['jahr_min'], Min('jahre__jahr'))
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('jahr_string', annotations)
         self.assertIsInstance(annotations['jahr_string'], Func)
 
@@ -694,7 +696,7 @@ class TestBaseBrochureAdmin(AdminTestCase):
         obj = (
             self.queryset
             .filter(pk=self.obj1.pk)
-            .annotate(**self.model_admin.get_changelist_annotations())
+            .annotate(**self.model.get_overview_annotations())
             .get()
         )
         self.assertEqual(self.model_admin.jahr_string(obj), '2001, 2002')
@@ -757,18 +759,17 @@ class TestBuchAdmin(AdminTestMethodsMixin, AdminTestCase):
         cls.test_data = [cls.obj1]  # noqa
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('autor_list', annotations)
-        self.assertIsInstance(annotations['autor_list'], ArrayAgg)
+        self.assertIsInstance(annotations['autor_list'], Func)
         self.assertIn('schlagwort_list', annotations)
-        self.assertIsInstance(annotations['schlagwort_list'], ArrayAgg)
+        self.assertIsInstance(annotations['schlagwort_list'], Func)
         self.assertIn('genre_list', annotations)
-        self.assertIsInstance(annotations['genre_list'], ArrayAgg)
-        self.assertIn('musiker_list', annotations)
-        self.assertIsInstance(annotations['musiker_list'], ArrayAgg)
-        self.assertIn('band_list', annotations)
-        self.assertIsInstance(annotations['band_list'], ArrayAgg)
+        self.assertIsInstance(annotations['genre_list'], Func)
+        self.assertIn('kuenstler_string', annotations)
+        self.assertIsInstance(annotations['kuenstler_string'], Func)
 
     def test_autoren_string(self):
         obj = self.get_annotated_model_obj(self.obj1)
@@ -862,10 +863,11 @@ class TestFotoAdmin(AdminTestMethodsMixin, AdminTestCase):
         cls.obj1 = make(cls.model, datum='2022-10-17', schlagwort__schlagwort=['Foo', 'Bar'])
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('schlagwort_list', annotations)
-        self.assertIsInstance(annotations['schlagwort_list'], ArrayAgg)
+        self.assertIsInstance(annotations['schlagwort_list'], Func)
 
     def test_foto_id(self):
         """Assert that foto_id returns the id padded with zeroes."""
@@ -891,10 +893,11 @@ class TestGenreAdmin(AdminTestMethodsMixin, AdminTestCase):
         cls.obj1 = make(cls.model, genre='Sub Genre', genrealias__alias=['Alias1', 'Alias2'])
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('alias_list', annotations)
-        self.assertIsInstance(annotations['alias_list'], ArrayAgg)
+        self.assertIsInstance(annotations['alias_list'], Func)
 
     def test_alias_string(self):
         obj = self.get_annotated_model_obj(self.obj1)
@@ -1044,10 +1047,11 @@ class TestMagazinAdmin(AdminTestMethodsMixin, AdminTestCase):
         )
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('orte_list', annotations)
-        self.assertIsInstance(annotations['orte_list'], ArrayAgg)
+        self.assertIsInstance(annotations['orte_list'], Func)
         self.assertIn('anz_ausgaben', annotations)
         self.assertIsInstance(annotations['anz_ausgaben'], Count)
 
@@ -1130,14 +1134,15 @@ class TestMusikerAdmin(AdminTestMethodsMixin, AdminTestCase):
         )
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('band_list', annotations)
-        self.assertIsInstance(annotations['band_list'], ArrayAgg)
+        self.assertIsInstance(annotations['band_list'], Func)
         self.assertIn('genre_list', annotations)
-        self.assertIsInstance(annotations['genre_list'], ArrayAgg)
+        self.assertIsInstance(annotations['genre_list'], Func)
         self.assertIn('orte_list', annotations)
-        self.assertIsInstance(annotations['orte_list'], ArrayAgg)
+        self.assertIsInstance(annotations['orte_list'], Func)
 
     def test_band_string(self):
         obj = self.get_annotated_model_obj(self.obj2)
@@ -1188,19 +1193,20 @@ class TestPersonAdmin(AdminTestMethodsMixin, AdminTestCase):
         cls.obj2 = make(cls.model)
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('is_musiker', annotations)
         self.assertIsInstance(annotations['is_musiker'], Exists)
         self.assertIn('is_autor', annotations)
         self.assertIsInstance(annotations['is_autor'], Exists)
         self.assertIn('orte_list', annotations)
-        self.assertIsInstance(annotations['orte_list'], ArrayAgg)
+        self.assertIsInstance(annotations['orte_list'], Func)
 
     def test_is_musiker(self):
         obj1, obj2 = (
             self.queryset
-            .annotate(**self.model_admin.get_changelist_annotations())
+            .annotate(**self.model.get_overview_annotations())
             .filter(id__in=[self.obj1.pk, self.obj2.pk])
             .order_by('id')
         )
@@ -1210,7 +1216,7 @@ class TestPersonAdmin(AdminTestMethodsMixin, AdminTestCase):
     def test_is_autor(self):
         obj1, obj2 = (
             self.queryset
-            .annotate(**self.model_admin.get_changelist_annotations())
+            .annotate(**self.model.get_overview_annotations())
             .filter(id__in=[self.obj1.pk, self.obj2.pk])
             .order_by('id')
         )
@@ -1248,10 +1254,11 @@ class TestPlakatAdmin(AdminTestMethodsMixin, AdminTestCase):
         cls.obj1.veranstaltung.set([v1, v2])
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('veranstaltung_list', annotations)
-        self.assertIsInstance(annotations['veranstaltung_list'], ArrayAgg)
+        self.assertIsInstance(annotations['veranstaltung_list'], Func)
 
     def test_datum_localized(self):
         self.obj1.refresh_from_db()
@@ -1323,10 +1330,11 @@ class TestSchlagwortAdmin(AdminTestMethodsMixin, AdminTestCase):
         )
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
         self.assertIn('alias_list', annotations)
-        self.assertIsInstance(annotations['alias_list'], ArrayAgg)
+        self.assertIsInstance(annotations['alias_list'], Func)
 
     def test_alias_string(self):
         obj = self.get_annotated_model_obj(self.obj1)
@@ -1385,12 +1393,11 @@ class TestVeranstaltungAdmin(AdminTestMethodsMixin, AdminTestCase):
         )
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
-        self.assertIn('musiker_list', annotations)
-        self.assertIsInstance(annotations['musiker_list'], ArrayAgg)
-        self.assertIn('band_list', annotations)
-        self.assertIsInstance(annotations['band_list'], ArrayAgg)
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
+        self.assertIn('kuenstler_string', annotations)
+        self.assertIsInstance(annotations['kuenstler_string'], Func)
 
     def test_kuenstler_string(self):
         obj = self.get_annotated_model_obj(self.obj1)
@@ -1430,12 +1437,11 @@ class TestVideoAdmin(AdminTestMethodsMixin, AdminTestCase):
         )
         super().setUpTestData()
 
-    def test_get_changelist_annotations(self):
-        annotations = self.model_admin.get_changelist_annotations()
-        self.assertIn('musiker_list', annotations)
-        self.assertIsInstance(annotations['musiker_list'], ArrayAgg)
-        self.assertIn('band_list', annotations)
-        self.assertIsInstance(annotations['band_list'], ArrayAgg)
+    def test_get_queryset_contains_annotations(self):
+        """Assert that the queryset returned by get_queryset contains the expected annotations."""
+        annotations = self.model_admin.get_queryset(self.get_request()).query.annotations
+        self.assertIn('kuenstler_string', annotations)
+        self.assertIsInstance(annotations['kuenstler_string'], Func)
 
     def test_kuenstler_string(self):
         obj = self.get_annotated_model_obj(self.obj1)